Move _get_target_surface to surface module
